# Conflicts:
#	app/build.gradle
#	app/src/main/java/ru/thstdio/mypetopenweather/data/Repository.kt
#	app/src/main/java/ru/thstdio/mypetopenweather/data/convertor/placeEntitytoPlace.kt
#	app/src/main/java/ru/thstdio/mypetopenweather/data/convertor/weatherPredictResponceToPlace.kt
#	core-db/src/main/java/ru/thstdio/core_db/impl/data/AppDb.kt
#	core-db/src/main/java/ru/thstdio/core_db/impl/data/dao/WeatherDao.kt
#	core-db/src/main/java/ru/thstdio/core_db/impl/data/entity/WeatherDBO.kt

+++ app/src/main/java/ru/thstdio/mypetopenweather/framework/api/response/WeatherPredictResponse.kt
@@ -12,7 +12,7 @@ data class WeatherPredictResponse(
     @SerialName("visibility") val visibility: Int,
     @SerialName("wind") val wind: WindResponse,
     @SerialName("clouds") val clouds: CloudsResponse,
-    @SerialName("dt") val dt: Int,
+    @SerialName("dt") val dt: Long,
     @SerialName("sys") val sys: SysResponse,
     @SerialName("timezone") val timezone: Int,
     @SerialName("id") val id: Long,
